refactor: migrate to new authentication hook - Step 9
- Replace useCanvasAuth with new useAuth hook in dashboard and sidebar
- Use isAuthenticated from lib/api/client in layout guard
- Simplify dashboard quiz fetching with useUserQuizzes hook
- Maintain existing functionality while improving code organization
